Split config layer utilities
